For #38944, Initial code reorganization
