+++ python/tk_multi_launchapp/__init__.py
@@ -1,0 +1,1 @@
+from launch_apps import InitAndLaunchApps
